src/requirements.txt
numpy==1.16.2
imageio==2.5.0
Pillow==6.2.0
scipy==1.3.1
PyQt5==5.12.2
dataclasses==0.6
pyyaml==5.1.2
psutil
